requirements-test.txt
-r requirements.txt
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
selenium==4.15.2
coverage==7.3.2
